+++ tests/testthat/test-tar_workspaces.R
@@ -1,7 +1,7 @@
 tar_test("tar_workspaces()", {
   expect_equal(tar_workspaces(), character(0))
   tar_script({
-    tar_option_set(workspace = c("x", "y"))
+    tar_option_set(workspaces = c("x", "y"))
     tar_pipeline(tar_target(x, "value"), tar_target(y, "value"))
   })
   tar_make(callr_function = NULL)